requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.2

# OpenAI API (更新版本)
openai==1.51.2

# JSON處理和數據驗證
pydantic==2.5.0

# 日誌記錄
structlog==23.2.0

# 開發工具
gunicorn==21.2.0

# HTTP 客戶端 (確保相容性)
httpx==0.27.0
